pyproject.toml
[tool.poetry]
name = "pypinger"
version = "0.1.0"
description = ""
authors = ["funnemonkee <ghostsmans@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "^6.0.2"
matplotlib = "^3.9.3"
pythonping = "^1.1.4"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
